implement MVP service macro
