+++ host-macro/Cargo.toml
@@ -13,7 +13,8 @@ quote = "1.0.7"
 darling = "0.20.10"
 proc-macro2 = "1.0.18"
 Inflector = "0.11.4"
-uuid = "1.2.2"
+
+uuid = "1.10.0"
 
 [lib]
 proc-macro = true
